FillRightCol, small bugfixes
Отлаживать и отлаживать.
И еще заменить функцию диффузии от стенки.


Former-commit-id: ff4ce6f6e74ac2a697136d478f54ee10bedba87e

+++ core/source/space.cpp
@@ -116,7 +116,7 @@ int Space::Print(TList<TObject> *List, const char* format)
 	fstream fout;
 	char fname[64];
 
-	sprintf(fname, format, int(Time/dt));
+	sprintf(fname, format, Time/dt);
 
 	fout.open(fname, ios::out);
 	res = Print(List, fout);
